requirements.txt
streamlit
python-dotenv
openai
python-docx

